Optimized Line-Triangle intersection with Moller-Trumbore

+++ Tactics/Engine/include/Engine/Triangle.h
@@ -25,7 +25,7 @@ namespace Engine
         // NaN if line does not intersect the triangle
         // positive distance from line.a to the interseciont if line.a is above the plane of the triangle
         // negative distance from line.a to the intersection if line.a is below the plane of the triangle
-        float Intersection(const Line& line) const;
+        double Intersection(const Line& line) const;
         Vector BaryCentric(const Coordinate& c) const;
     };
